% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tile.R
\name{get.tile}
\alias{get.tile}
\title{Gets Mapbox Vector Tile row and column}
\usage{
get.tile(zoom, lat, long)
}
\arguments{
\item{zoom}{zoom level}

\item{lat, long}{latitude and longitude of target point}

\item{long}{}
}
\value{
rowcol      two element vector with row and column in MVT
}
\description{
Translates latitude and longitude at a given zoom level to row and column of a
Mapbox Vector Tile.
}
\details{
Mapbox Vector Tiles use a common addressing scheme shared with other web mapping
services, such as Open Street Map Slippy Tiles. Tiles numbers are the same for any
data source, based only on zoom level, latitude, and longitude. This function returns
the tile row and column.
Translated from Python and JavaScript, thanks to
    \url{https://stackoverflow.com/questions/29218920/how-to-find-out-map-tile-coordinates-from-latitude-and-longitude}
}
\examples{
require readMVT
xml <- read.XML('https://umassdsl.webgis1.com/geoserver')
info <- layer.info(xml, 'testbed:streamlines')
rc <- get.tile(10, 48.0096, -88.7712)

}
