% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tile.R
\name{read.tile}
\alias{read.tile}
\title{Read Mapbox Vector Tile from a GeoServer}
\usage{
read.tile(info, zoom, row, col)
}
\arguments{
\item{info}{layer info for target layer, from \code{\link[=layer.info]{layer.info()}}}

\item{zoom}{zoom level}

\item{row, col}{tile row and column (from \code{\link[=get.tile]{get.tile()}})}
}
\value{
sf object
}
\description{
Read Mapbox Vector Tile for layer described by info, specified zoom, row, and column
from a GeoServer and return as an sf object.
}
\details{
This uses the URL in the MVT object created by layer.info to produce a URL for
the specified row and column at the specified zoom level to fetch the MVT data
from a GeoServer and return the data as an sf object.

Note that this functions requires protolite >= 2.2.1. If it's not yet up on CRAN,
you can get it with

\code{remotes::install_github('jeroen/protolite')}
}
\section{Author}{

Bradley W. Compton \href{mailto:bcompton@umass.edu}{bcompton@umass.edu}
}

\examples{
require(readMVT)
require(leaflet)
xml <- read.XML('https://umassdsl.webgis1.com/geoserver')
info <- layer.info(xml, 'testbed:streamlines')
rc <- get.tile(12, 42.394, -72.5312)
x <- read.tile(info, 12, rc[1], rc[2])
leaflet(x) |> addTiles() |> addPolylines(data = x)

}
