% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.info.R
\name{layer.info}
\alias{layer.info}
\title{Read info for a Mapbox Vector Tile layer on a GeoServer}
\usage{
layer.info(xml, layer, crs = "EPSG:900913")
}
\arguments{
\item{xml}{XML from getcapabilities. Use \code{\link[=readXML]{readXML()}} to get this}

\item{layer}{name of the layer, e.g., testbed:streamlines}

\item{crs}{coordinate reference system. For Mapbox Vector Tiles, you'll want EPSG:900913 (= EPSG:3857)}
}
\value{
An MVT object, a four element list with:
\enumerate{
\item layer: name of layer (colon converted to underscore)
\item box: bounding box (x-min, y-min, x-max, y-max)
\item tiles: n x 5 matrix of zoom level, rowmin, rowmax, colmin, colmax
\item url: constructed URL, with {style} and {TileMatrixSet} replaced, and {TileMatrix} replaced with {zoom}
leaves {zoom}, {TileRow}, and {TileCol} to be replaced on reads
}
}
\description{
Reads all necessary info for reading Mapbox Vector Tiles from a GeoServer.
}
\details{
Pulls necessary information for reading Mapbox Vector Tiles from a GeoServer's capabilities XML. Builds
a template URL, with tags to be replaced by read.tile, when the target zoom level, row, and column are
known.
}
\section{Author}{

Bradley W. Compton \href{mailto:bcompton@umass.edu}{bcompton@umass.edu}
}

\examples{
require(readMVT)
xml <- read.XML('https://umassdsl.webgis1.com/geoserver')
info <- layer.info(xml, 'testbed:streamlines')

}
